Downgrades `package-lock.json` for NPM 6.
Apparently NPM 7 does not consistently install unsupported optional deps on `npm ci`, which was necessary to workaround this bug: https://github.com/bazelbuild/rules_nodejs/issues/2395

As a result, we can't really use NPM 7 right now. I also added docs to prefer the Bazel version of NPM and Node to be more consistent about this. Also pinned the relevant versions in `package.json` so users are less likely to make this mistake.