Adds a browse test for a prerendered counter component.
Refs #36.

This prerenders a `site-counter` component and then loads it in Puppeteer based the second strategy suggested in https://github.com/dgp1130/rules_prerender/issues/36#issue-853048853.

This requires quite a bit of boilerplate but is relatively intuitive and mostly works out of the box. It does rely on testing helpers private to `rules_prerender` which real users would not be able to take advantage of (`useDevserver()`, `usePuppeteer()`, `resolveRunfile()`). Regardless, it tests the critical integration point of JavaScript running on prerendered HTML, verifying that the content is hydrated and used correctly.

This does include the known downside of using separate files for generating test cases and actually running the tests. It requires a bit of unfortunate back and forth, but isn't too bad for a simple counter at least. Some helpers could be introduced to reduce the boilerplate, such as a `generateTestCase('Zero', () => `<site-counter>...</site-counter>`)` which could generate all the HTML scaffolding for the page with a given title and body content. `useDevserver()` and `usePuppeteer()` could be turned into public APIs (maybe less opinionated though).

One unfortunate limitation of the current approach is that the page hydrates too fast for Puppeteer to inspect the browser DOM. We could drop the prerender unit test and merge it into this if it were possible teo inspect the component DOM before it hydrates. It would allow us to easily test that the component prerendered to a good state (initial value visible, buttons disabled, etc.) before testing the loaded JavaScript. Again, some more opinionated test helpers might be able to defer JS execution to make this possible.