Adds `web_resources()` validation action.
Closes #22.

This uses the experimental validation action to make each `web_resources()` target assert that it is able to merge all its entries together. In normal execution, only the top-level `web_resources()` target actually merges its entries, all others are skipped. However, this means that any merge conflicts are surfaced at the top-level `web_resources()` target, long after the relevant mistake was actually made. This validation action makes the offending `web_resources()` target fail, putting the error message at the location of the user's mistake.

This unfortunately requires `--experimental_run_validations` to be executed, but it's pretty easy to add that to the `.bazelrc` file.

I updated the progress messages to include the target being compiled. Apparently this information is not included by default, which I found particular confusing during debugging. It's good practice to include this for all actions.

I also included a test case to confirm that the validation action actually works. Unfortunately I could not find a good way to actually convert a failing target into a passing test. Starlark does support [analysis failure tests](https://docs.bazel.build/versions/4.1.0/skylark/testing.html#failure-testing), but I believe this is only capable of verifying that `fail()` was called correctly, not that a specific action failed.

Ultimately, the best I could do was to make the test case and put `tags = ["manual"]` on the failing targets. This means that `bazel build //...` will still pass because manual targets are not run, but `bazel build //path/to/pkg:target` will fail with the relevant error, serving as an easy reproduction. Unfortunately, I don't see any way to actually run this with a `bazel test //...`, so this is *not* run automatically via CI.