Adds `inlineStyle()` function.
Refs #38.

The `inlineStyle()` function simply loads the given input from runfiles and returns it in a `<style />` element. This is useful for inlining styles in specific parts of the document, such as inside a Declarative Shadow DOM template.

Since this function directly reads the CSS file and returns its contents, it must be `async`, which can be awkward an unintuitive to use. I expect most users will do `<div>${inlineStyle('wksp/foo.css')}</div>`, but this would actually print `<div>[object Promise]</div>` becuase the user didn't `await`. Unfortunately neither TypeScript nor ESLint seem to catch this mistake out of the box, and I expect it will be a common problem. For now, this is good enough to unblock Declarative Shadow DOM, however in the future it would be a good idea to make this synchronous and leave a style annotation which is processed later to actually inline the style. Doing so leads to a bunch of other questions that I don't want to get into right now.

1. Is inlining done by the annotation extractor or resource injector?
2. If done by the resource injector, how do we keep things deterministic? Is it weird that annotation extractor actually leaves annotations for a future tool to process?
3. How do we enforce strict deps?
4. Should we allow users to resolve package-local files?