Adds JS-only example.
Refs #39.

This example application only uses `*.js` source files to validate the use case. It includes JavaScript prerendering code, both in a `prerender_page()` and a standalone `prerender_component()`. It also includes JS client-side scripts and validates that they get included in the final bundle *and* are properly tree shaken when not used. This also includes a dependency on another client-side `js_library()` which is defined with ESM and bundled via Rollup.