Adds `//common:paths.bzl`.
Refs #39.

This is a simple Starlark library for managing paths. For now it just has an `is_js_file()` function to return whether or not a file is JavaScript based on its extension. There are actually a few different possibilites so this is not quite trivial and I don't want to repeat this everywhere it is necessary.