Ignores `user/` directory.
This is used by the `ref/external` branch for a test user repository. When checking out from that branch, the files are considered untracked in the new branch and often left as is, where it will break Bazel. By ignoring this directory in the `main` branch, we ensure that checking out from `ref/external` won't break anything.