Adds `pre-push` lint hook to regret later.
This adds `husky` to the project and introduces a `pre-push` Git hook which lints all the commits being pushed to `main`. If any commit fails linting, the push is aborted. This should be useful since most `main` commits are pushed directly rather than via a PR where a GitHub action would run.

Unfortunately this is entirely Linux (and maybe Mac?) specific, this would not work for Windows. I don't yet support that as a dev machine anyways though, so I guess it doesn't matter for now, but it will be something to solve later on.

This ended up a lot more complicated than expected and I'm sure I will regret this eventually. I didn't want to use `pre-commit` hooks because they often drop the commit message and I don't expect every commit to pass linting, such as WIP commits. On push to `main` seemed more appropriate because that is when valid linting should be required.