Adds `prerender_component_publish_files()`.
Refs #39.

This macro collects all the files generated by a `prerender_component()` target so they can be included in a published NPM package. This allows a component to be written in TypeScript and used normally within its source workspace, but then have its generated artifacts (`.js` and `.d.ts` in particular) published to an external package manager (typically NPM). Since we already support `prerender_component()` using `*.js` and `*.d.ts` sources, the published content can include its own, separate build file which makes its own `prerender_component()` that uses those generated artifacts. That component can then be depended upon by users of the package as if it were a regular `prerender_component()` compiled directly from source.

By doing this we don't ship TypeScript to NPM and we decouple the source files that actually generate a `prerender_component()` from the published files that end up on users machines. In particular, this means that we don't require a TypeScript compiler for the user, we don't need a `tsconfig` and we don't need to worry about compiler versioning for every build tool we use.

I debated a lot about whether or not we should include `*.d.ts` files for scripts, since they won't actually be type checked anywhere. However, I ultimately decided to include them because they can have useful information, there's not much harm in propagating them, and future or alternative bundling tools might take advantage of them. Understanding *why* certain `*.js` files don't have associated `*.d.ts` files would also be quite confusing and hard to trace back, so I think it's better to avoid that nuance to begin with.

It's tricky to automatically test this without a complex integration testing setup using multiple workspaces and passing through a `pkg_npm()` target. For now, we just assert that the files we expect are present in the output, which is good enough for the time being.