Updates `prerender_component_publish_files()` to include transitive source files.
Refs #39.

This is necessary to include transitive dependencies of the prerender library and the script libraries in the published package. It also includes transitive prerender and script declarations, which aren't strictly necessary but stil useful for debugging purposes and to avoid the inevitable question "Why do some `*.js` files not have associated `*.d.ts` files?".