Adds resources to published `prerender_component()`.
Refs #39.

This requires specifying the path mapping twice, once in the `BUILD.bazel` and again in the published `BUILD` file. This is kind of annoying but I don't have a better way of doing this atm.

It also leaks the implementation detail of the `%{name}_resources` target, since it gets included directly in the output package. I think this is ok since there is no better format it could be in (can't easily map a root-scoped `web_resources()` directory to a package-scoped component, multiple components might conflict, etc). This should be good enough for the time being at least.