Removes `@${COMMIT}` from `pre-push` script and adds `--report-unused-disable-directives`.
The `--stdin-filename` flag is not just for display purposes, it is also used to determine which rules should be run for the particular file. By including the hash at the end, the files were not being correctly detected as TypeScript and using the wrong settings. This meant that `git push` and `npm run lint` disagreed, which is particular confusing.

Workaround is to pass the filename in unchanged and update the `pre-push` script to print which commit failed linting. This is a bit awkward because it must save the output of the linter into a variable and re-print it only if linting failed, however doing so would drop color information. The `script` command apparently preserves this info, which keeps the output from degrading.