Updates `site` example's counter component to use declarative shadow DOM.
This required some tweaks in the client JS to read from nodes in the shadow DOM. It assumes that the `shadowRoot` will be present since we know the component must have been prerendered with a declarative shadow DOM root.

This also required adjusting the tests to get nodes from the shadow root. For this test I opted to use `evaluateHandle()` to get the shadow root and `$eval()` from there. Alternatively we could use `$eval('pierce/.some-selector')`, but querying that shadow root directly is less dependent on the rest of the DOM tree IMHO.