Updates CI to not abort on certain failures and to only build targets actually needed by tests.
This makes CI show as many failures as it can rather than aborting because it encountered a Bazel issue in one package. It also avoids building targets not needed for tests, which is good for performance. If a target doesn't have a test but should be verified as buildable by CI, then it should have a `build_test()`.

See https://www.youtube.com/watch?v=j332WfhNAFg (approx. 22m40s)