Uses CI configuration for release workflow.
This speeds up the build/test cycle because it reuses the cache from previous builds. It also uses the CI configuration for the actual release run, though this has less caching benefit since everything is stamped. However, the CI configuration is still valuable due to the improved output formatting and profiling.