Updates CI to use Bazel-versioned NodeJS install.
Installing Node via GitHub duplicates the install and introduces a second version of the binary. NodeJS is already exposed in the workspace, so we can just run that directly, which will keep the version consistent. `bazelisk` is actually installed by default in the `ubuntu-20.04` image, so we can just use that to bootstrap the build.

This also caches the `node_modules/` directory, which seems to be necessary for Bazel to effectively used previously cached artifacts. I'm not totally sure why that's necessary, as I'm pretty sure `npm ci` deletes `node_modules/`, but if we don't do this Bazel doesn't seem to use any caching.