Moves `--define VERBOSE_LOGS=1` from `--config debug` into it's own `--config debug-resolve`.
This is only necessary when debugging Node resolution issues, which is a small fraction of debugging use cases. Otherwise this is just noise in the logs. Since it is not frequently used, `--config debug` no longer includes it, and users must explicitly opt-in to logging Bazel Node resolution via `--config debug-resolve`.

I considered dropping this flag altogether since `--config debug-resolve` isn't any more convenient than `--define VERBOSE_LOGS=1`, but I think it is useful to exist in `.bazelrc` to be more discoverable. Finding this flag in `rules_nodejs` can be quick tricky, but skimming the `.bazelrc` for useful debug flags would quickly find it.