Adds WebDriverIO and an effect to use it.
This installs `webdriverio` and uses it to drive a Chrome browser and assert on a web page. Most WebDriverIO tools and dropped and not used here since we don't run WebDriverIO as a tool and don't need its Jasmine or TypeScript integrations. Instead we just treat the core `webdriverio` as a library and use it in a `jasmine_web_test_suite()` target, which sets up the WebDriver server and sets its URL in an environment variable. The WebDriverIO testing library reads this environment variable and connects to it to start a browser.

The effect itself has no cleanup action. WebDriverIO examples include `await browser.deleteSession();` at the end, but doing this here causes an error due to deleting a non-existent section. I think something is automatically cleaning up the browser, so manually deleting the session results in a double-free error. I also updated `useDevserver()` to be usable as an input to `useWebDriver()`, since that will be quite common.

This unfortunately *requires* `types: ["webdriverio/async"]` in the `tsconfig.json`. I would much rather directly import all the required types, but that doesn't seem possible. Even if I directly import the type I want (such as `Browser`), it is dependent on other global types (such as `WebDriverIO.Element`) and they get converted to `any`, making these types much less useful. Rather than trying to fight it, I've just accepted that the global types are needed.

Includes documentation about WebDriverIO and the `--config debug` flag. Left in the Puppeteer stuff for now until it is completely removed.