Migrates Declarative Shadow DOM test from Puppeteer to WebDriverIO.
I was worried that Declarative Shadow DOM might not be fully supported or hidden behind a flag, but it actually worked quite well out of the box. WebDriverIO even supports a `.shadow$()` method which is much more convenient than Puppeteer.

Ironically, the harder part was `getComputedStyle()`, which apparently isn't supported by WebDriver. Instead I had to `.execute()` some JS to get the color, which adds a bit of complexity. This will probably get pulled out into a helper function at some point, but for now I'll just leave it in this test.