Migrates custom bundling example from Puppeteer to WebDriverIO.
Pretty standard. Only unique aspect is `.getPageSource()` to look for unprocessed annotations. This technically returns a string representation of the current DOM, not the original page source it loaded with. This makes the assertion slightly wrong, as the runtime JS could have replaced an annotation that should not have been served to begin with. I'm too lazy to work around it though, and I don't think it's a major issue.