Moves `getColor()` into the WebDriver helper utility.
This function is useful outside of just Declarative Shadow DOM and is helpful for asserting styles in other pages (which are typically expressed as color differences). It accepts an element, or a `ChainablePromiseElement` to support the asynchronous syntax typically used with WebDriverIO. I also needed to accept `ChainablePromiseElement<Promise<WebdriverIO.Element>>`, which seems redundant to me, but is the return type of `browser.$('selector')`, so it is necessary.