Adds `scope` to CSS prerender annotations.
Refs #41.

Scope refers to how the CSS file is used. `Global` means that it applies to the entire page, while `Inline` means it is inlined into a particular location in the document and limited to any shadow DOM it might be under. Currently `includeStyle()` is considered global, while the new `inlineStyle()` will be inlined in its specified location.