Adds `prerender_annotation_walker`.
This presents a nice API for iterating over annotation comments in a DOM tree with the ability to remove and replace them. Normally you need to track a node's parent to be able to remove or replace it, so this abstracts away that process and gives a single `UpdateableNode` class which can be removed / replaced directly without thinking about the parent.