Fix VSCode debugger to resolve sourcemaps correctly.
Using the debugger would open temporary files not in the source tree. After fixing this it is able to resolve them to their original source files and reduce the number of tabs to worry about by eliminating duplicate files.

TBH I don't really understand what this change does, I just looked at the linked doc and updated the configuration to match. I'm not sure if it changed since setting this up, or exactly what happened here.