Removes style entry point from `prerender_pages_unbundled()` and public API.
Refs #47.

This is no longer used by `prerender_pages()` and I don't want to support it just for custom bundling use cases. Without `includeStyle()` there is no way to identify which CSS files are associated with which page, so the whole system couldn't be optimized enough to be worth using. Instead I'm tearing down the whole thing and recommending inline styles until someone gives a valid use case for `includeStyle()`.