Removes `styles` example.
Refs #47.

This was originally intended to test global styles, but it has since been updated to test inline styles despite a separate `//examples/inline_styles/...` package which serves that purpose better because it includes CSS imports and dependencies. As a result, the `styles` example is no longer providing any meaningful value.