Switches all tools to `host` configuration.
Refs #50.

`exec` is generally preferred, however it is not easily supported by `ts_project()` because we need to explicitly list all the Bazel artifact roots in `rootDirs` (https://bazelbuild.github.io/rules_nodejs/TypeScript.html#ts_project). `exec` dynamically computes artifact roots based on the target's configuration, which can change arbitrarily based on build flags and how targets are used. It's not fit to be hard-coded in a `tsconfig.json`. I left a `TODO` to switch back to `exec` after updating to `rules_js`, since it should be able to fix this particular rough edge.