Migrates to use `jasmine_test()` from `@aspect_rules_jasmine`.
Aspect now provides a real Jasmine ruleset for `@aspect_rules_js`. This drops some of my bespoke implementation and instead uses theirs. This still requires a custom config to be used because otherwise Jasmine is restricted to the default file conventions. I decided to leave in my rule which generates a precise config exactly referencing direct dependencies as files to test, this way there are no magic naming conventions.