Migrates renderer to `js_binary()` from `@aspect_rules_js`.
This also necessitated updating `prerender_component()` and some examples to use the `@aspect_rules_js` implementation of `js_library()`.

NOTE: This commit does not pass CI but will be fixed in a later commit. In particular, `ts_project()` is still using the `@build_bazel_rules_nodejs` implementation which is not compatible and can't be easily landed incrementally. Instead I'm deliberately accepting CI breakage for the next few commits as the renderer is migrated to `@aspect_rules_js` and the whole repository is migrated to `@aspect_rules_ts`.