Updates renderer to use a generated entry point instead of dynamic importing.
Refs #48.

This simplifies file resolution since it comes from the main file in the same package generating the `prerender_resources()` target as opposed to coming from `packages/renderer/dynamic_import.js`. This makes it easier to port to `@aspect_rules_js`.

I restructured the test to run in-process instead of calling the tool via a subprocess. This is because the real entry point gets generated by `prerender_resources()`, so the tool can't be invoked independently. This has the side effect of making the linking process much simpler, since we don't have to generate `*.js` files which get executed and need to `require()` things.

Unfortunately, for some reason the `rules_prerender` import isn't working right now, but does when migrated to `@aspect_rules_js`. So I disabled the test for now and will re-enable it once the renderer is migrated to `@aspect_rules_js`. I'm not sure what's up with the import, but since we're moving off this infrastructure anyways, I don't think it matters very much.