Migrates `ts_config()` targets to use the `@aspect_rules_js` implementation.
This means the base tsconfig is now its own target and should be referenced as such. I remember trying to do this with `@build_bazel_rules_nodejs` but failing, so I'm glad this is now more aligned with the way I expected the dependency structure to work. Unfortunately it means I had to update a number of `//:tsconfig.json` references to be plain `//:tsconfig` references.

NOTE: Tests are still broken at this commit.