Updates `includeScript()` usages to drop workspace name.
`@aspect_rules_rollup` configuration requires dropping the workspace name so everything is imported relatively. This means it is an open question of exactly how to import scripts from other workspaces. It also introduces divergence with the way `inlineStyle()` works, but both of those are future problems.

NOTE: Tests are still failing at this commit.