Adds `link_prerender_component()` with support for prerender and script JS.
This does not implement style or resource files because those are not strictly necessary for declarative shadow DOM, which doesn't use either. The current implementation also mixes prerender and script files together, meaning you could `includeScript()` a prerender file or `import` a script file, neither of which is desirable. After DSD is mostly functional, we'll have to revisit a more comprehensive solution here which splits out the NPM package into multiple `TreeArtifacts`.

NOTE: Tests are still failing at this commit.