Updates `prerender_component_publish_files()` to use `JsInfo`.
This updates the macro to be compatible with `@aspect_rules_js`. I tried making it compatible with both by supporting `JsInfo`, `JSModuleInfo`, and `DeclarationInfo`, however these providers are not mutually exclusive. `ts_project()` from `@aspect_rules_js` appears to return all three and exactly how to handle that behavior is not obvious. Instead of trying to fight it, I just decided to make this a one-time fix.

NOTE: Tests still do not pass at this commit.