Migrates all `concatjs_devserver()` usages to `js_run_devserver()`.
Refs #48.

I mostly followed instructions in [Aspect docs](https://docs-legacy.aspect.build/aspect-build/rules_js/v1.6.4/docs/js_run_devserver-docgen.html#js_run_devserver). They happened to use `http-server` as the server dependency, which is also what I was familiar with.

This implementation encountered a couple interesting challenges along the way:

First, the devserver needed to implicitly add the served directory as an argument, so the actual command line would be `http-server path/to/my/resources/`. This could be trivially done with `js_run_devserver(args = ["path/to/my/resources/"])`, however this was implemented via Bazel's built-in argument handling logic for executable rules, which has the behavior that the args are more like "default args". They can be overridden by any program invoke the executable, whether on command line like `bazel run //:devserver -- some args` or when executing the server as a data dependency of a separate target like the e2e tests. This meant that the devserver "forgot" what directory it was supposed to serve. I solved this via a `_baked_binary()` target, which generates a shell script which includes the resource directory and any other arguments append to instead of overwrite the existing arguments.

This ran into the second problem, where `js_run_devserver()` assumes that the underlying target is a `js_binary()` and does `process.chdir()` into the output directory before executing the subprocess. This wouldn't be that big a deal, but the baked binary needs to execute the `http-server` binary under the hood, which is stored in its runfiles. This means we technically needed [`@bazel_tools` Bash runfiles helper](https://github.com/bazelbuild/bazel/blob/668b0da527a43299f6f6ac49bb5cc37be2265c45/tools/bash/runfiles/runfiles.bash) to find this binary. However that doesn't actually work because the baked binary is in a different working directory than expected and fails to find its runfiles. The easiest solution is to use the `$RUNFILES` environment variable, though TBH I'm not 100% on where that is coming from or how consistently it is set to what I want. Seems to work for now so 🤷‍♂️. This also means that we can't make a generic implementation of `_baked_args()` for re-use, since a generic version wouldn't be easily compatible with `js_run_devserver()` without including a modified version of the Bash runfiles helper, which I don't want to get into right now.

The third problem is that `js_run_devserver()` leaks the fact that it is implemented as a shell script under the hood. You can't actually invoke it by its name, but instead need to append `.sh` to the end, which probably only works for Linux hosts. That's fine for now, but might be an issue for cross-compilation later.

The fourth problem is that there are a couple one-off `concatjs_devserver()` usages for testing test utilities. I wanted to migrate these to direct usages of `js_run_devserver()` with one-off `http-server` binaries, but they also run into the "forgetting the served directory" problem and need the baked args implementation. Instead of trying to share that, I just decided to reuse `web_resources_devserver()`. This is bending dependencies in a slightly undesriable manner, since test utilities shouldn't have anything to do with `web_resources_devserver()` in particular or any other `rules_prerender` packages/tools. But this is the easiest way to get those tests working, and it's not worth the effort to fight it.

One other challenge I'm not dealing with right now is the fact that because `js_run_devserer()` propagates arguments, this actually leaks the underlying `http-server` implementation detail. Users could `bazel run //:some_web_resources_devserver -- some http-server args` and it would actually work. This would be bad for abstraction and makes it harder to change the implementation in the future. Instead, the only flag we really want/need to support is `--port` (and maybe `--address`). The rest should be actively blocked to prevent users from accidentally depending on it in weird ways.