Adds qualified `@rules_prerender` workspace name to absolute labels used in macros evaluated in user workspaces.
Refs #48.

Macros evaluated in user workspaces interpret absolute labels of the format `//foo/bar:baz` as being contained in the user's workspace. We need to qualify them with `@rules_prerender//foo/bar:baz` to specifically refer to the `@rules_prerender` workspace. This doesn't break local development in `@rules_prerender`, because that is still a valid label to the main workspace.