Splits `package.bzl` to work with `@aspect_rules_js`.
Refs #48.

Bazel restricts `load()` statements to only be used in a `WORKSPACE` file, `BUILD` file, or the root scope of a `*.bzl` file. This means we can't write workspace macros like:

```starlark
def dependencies():
    http_archive(name = "foo", ...)
    load("@foo//:defs.bzl", "bar") # ERROR: Cannot use `load()` outside root scope.
    bar()
```

This makes supporting workspace files a bit awkward, since we need to push all the `load()` statements into user-space. Many other `@aspect_rules_js` rulesets use a similar model of `dependencies()`, `repositories()`, and `npm_repositories()`, and this seems good enough for what `@rules_prerender` needs right now.

Note several repositories are included as dependencies here even when they really shouldn't be. For example, Jasmine and `@rules_webtesting` should not be leaked into user code and should never be built, however they are used by targets which share build files with user code, so their definitions must be loaded, even if they are never built. This is not ideal, and we should revisit this in the future, but for now just adding the dependency is good enough for things to build.