Injects user-space `rules_prerender` dependency into renderer.
Refs #48.

We need to use the `rules_prerender` package linked in the root package of the user's workspace, not the one built in `@rules_prerender`. This pushes the dependency into the generated `js_binary()` target which is created in the user's workspace, not `@rules_prerender`. This means it can have a dependency on the `//:node_modules/rules_prerender` target in the same workspace, which will align with the one used in user-space and avoid duplicating the dependency's implementation.

This does assume that the `rules_prerender` package is linked at `//:node_modules/rules_prerender`, which is _probably_ the case, but might not. We should probably add an additional attribute for `rules_prerender_dep` to all the macros to allow it to be configured.