Upgrades GitHub actions.
This upgrades the actions to v3 which supports Node versions greater than 12. The "Actions" page was showing deprecation warnings and this alleviates them.

I'll confess I didn't actually go through the breaking changes on all these, but my usage is simple enough that it _hopefully_ won't be a problem. I confirmed that CI still passed, but the lint and release actions are harder to test in advance.